+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -153,6 +153,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
     }
 
     @Test
+    @TestMetadata("InvalidGenericScopeAliases.kt")
+    public void testInvalidGenericScopeAliases() {
+      run("InvalidGenericScopeAliases.kt");
+    }
+
+    @Test
     @TestMetadata("MissingBindingInContributionProviderRedirectsToOrigin.kt")
     public void testMissingBindingInContributionProviderRedirectsToOrigin() {
       run("MissingBindingInContributionProviderRedirectsToOrigin.kt");